AR-611: make themes-columns.jsx
